+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr2 (development version)
 
+* `curl_translate()` now works with multiline commands from the clipboard
+  (@mgirlich, #254).
+
 * New `resp_has_body()` returns a `TRUE` or `FALSE` depending on whether
   or not the response has a body (#205).